Serialization pipeline implemented

+++ SynapseEngine/Engine/Utils/RenderBuffer.cpp
@@ -38,7 +38,7 @@ namespace Syn
         }
     }
 
-    bool RenderBuffer::UpdateCapacityAll(uint32_t requiredElements)
+    bool RenderBuffer::UpdateCapacityAll(uint64_t requiredElements)
     {
         bool resized = false;
         for (uint32_t i = 0; i < _config.frames; ++i) {
@@ -49,7 +49,7 @@ namespace Syn
         return resized;
     }
 
-    bool RenderBuffer::UpdateCapacity(uint32_t frameIndex, uint32_t requiredElements)
+    bool RenderBuffer::UpdateCapacity(uint32_t frameIndex, uint64_t requiredElements)
     {
         bool resized = false;
         if (frameIndex < _config.frames) {
